reorganizing functions - just moving them aronud, no changes.
